ChunkStore code modularized with test. Helper module integrated. Unit tests pending
