+++ src/chunk_store/chunk_store.rs
@@ -141,61 +141,3 @@ impl ChunkStore {
   //   self.current_disk_usage + required_space <= self.max_disk_usage
   // }
 }
-
-mod test {
-    extern crate maidsafe_types;
-    extern crate rand;
-    use rand::{thread_rng, Rng};
-    use std::str;
-    use chunk_store::ChunkStore;
-
-    static ONE_KB: u64 = 1024u64;
-    static AES_PADDING: u64 = 16u64;
-    // Allow 16 bytes extra per chunk since we're AES encrypting them
-    static K_DEFAULT_MAX_DISK_USAGE: u64 = 4 * 1024u64 * 16u64; // 4 * (OneKB + AesPadding);
-
-    struct NameValueContainer(Vec<(maidsafe_types::NameType, String)>);
-
-    fn get_random_name_type() ->  maidsafe_types::NameType {
-        let mut v = [0u8; 64];
-        thread_rng().fill_bytes(&mut v);
-        maidsafe_types::NameType(v)
-    }
-
-    fn get_random_non_empty_string(length: usize) -> String {
-        let mut string = String::new();
-        for char in rand::thread_rng().gen_ascii_chars().take(length) {
-            string.push(char);
-        }
-        string
-    }
-
-    fn add_random_name_value_pairs(number: u32, size: u32) -> NameValueContainer {
-        let mut i = 0u32;
-        let mut container: Vec<(maidsafe_types::NameType, String)> = Vec::with_capacity(number as usize);
-        loop {
-            container.push((get_random_name_type(), get_random_non_empty_string(size as usize)));
-            i = i + 1; // i++; is not compiling
-            if i == number {
-                break;
-            }
-        }
-        NameValueContainer(container)
-    }
-
-    // TODO (Krishna) - Must support file storage, once the ChunkStore supports file storage. At present ChunkStore is in-memory storage
-    struct ChunkStoreTest {
-        chunk_store: ChunkStore,
-        max_disk_storage: u64
-    }
-
-    impl ChunkStoreTest {
-        pub fn new() -> ChunkStoreTest {
-            ChunkStoreTest {
-                chunk_store: ChunkStore::new(),
-                max_disk_storage: K_DEFAULT_MAX_DISK_USAGE
-            }
-        }
-    }
-
-}
