+++ src/chunk_store/mod.rs
@@ -1,0 +1,3 @@
+mod chunk_store;
+mod test;
+pub use self::chunk_store::*;
